+++ inject/inject-thrift-client/src/test/scala/com/twitter/inject/thrift/integration/basic/EchoThriftClientModule1.scala
@@ -1,7 +1,8 @@
 package com.twitter.inject.thrift.integration.basic
 
 import com.twitter.finagle.ThriftMux
-import com.twitter.inject.thrift.modules.{PossiblyRetryableExceptions, ThriftClientModule}
+import com.twitter.inject.exceptions.PossiblyRetryable
+import com.twitter.inject.thrift.modules.ThriftClientModule
 import com.twitter.test.thriftscala.EchoService
 import com.twitter.util.Future
 
@@ -13,6 +14,6 @@ object EchoThriftClientModule1 extends ThriftClientModule[EchoService[Future]] {
     client: ThriftMux.Client
   ): ThriftMux.Client = {
     client
-      .withResponseClassifier(PossiblyRetryableExceptions)
+      .withResponseClassifier(PossiblyRetryable.ResponseClassifier)
   }
 }