+++ inject/inject-thrift-client/src/test/scala/com/twitter/inject/thrift/integration/serviceperendpoint/GreeterThriftMethodBuilderClientModule.scala
@@ -4,9 +4,10 @@ import com.google.inject.Module
 import com.twitter.finagle.service.{ReqRep, ResponseClass, ResponseClassifier}
 import com.twitter.greeter.thriftscala.Greeter.Bye
 import com.twitter.greeter.thriftscala.{Greeter, InvalidOperation}
+import com.twitter.inject.exceptions.PossiblyRetryable
 import com.twitter.inject.thrift.ThriftMethodBuilderFactory
 import com.twitter.inject.thrift.integration.filters.{HiLoggingTypeAgnosticFilter, MethodLoggingTypeAgnosticFilter}
-import com.twitter.inject.thrift.modules.{PossiblyRetryableExceptions, ThriftClientIdModule, ThriftMethodBuilderClientModule}
+import com.twitter.inject.thrift.modules.{ThriftClientIdModule, ThriftMethodBuilderClientModule}
 import com.twitter.util.{Return, Throw}
 import scala.util.control.NonFatal
 
@@ -28,7 +29,7 @@ object GreeterThriftMethodBuilderClientModule
         builder.method[Greeter.Hi.Args, Greeter.Hi.SuccessType](Greeter.Hi)
           // method type-agnostic filter
           .withAgnosticFilter[HiLoggingTypeAgnosticFilter]
-          .withRetryForClassifier(PossiblyRetryableExceptions)
+          .withRetryForClassifier(PossiblyRetryable.ResponseClassifier)
           .service)
       .withHello(
         builder.method(Greeter.Hello)
@@ -40,7 +41,7 @@ object GreeterThriftMethodBuilderClientModule
         builder.method[Bye.Args, Bye.SuccessType](Greeter.Bye)
           // method type-specific filter
           .filtered[ByeFilter]
-          .withRetryForClassifier(PossiblyRetryableExceptions)
+          .withRetryForClassifier(PossiblyRetryable.ResponseClassifier)
           .service)
       // global (type-agnostic) filter
       .filtered(new MethodLoggingTypeAgnosticFilter())